Web client compilation * Removed web specific files. Now each module that needs to have different code for web will do it itself * new Util.isBrowser() function to determine if the sdk is running in a browser or not * Now that transport layer has been abstracted away (transport.js), can remove special code to trace SOAP calls when running in browser
