--- test/web.crypto.test.js
@@ -1,30 +1,0 @@
-/*
-Copyright 2020 Adobe. All rights reserved.
-This file is licensed to you under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License. You may obtain a copy
-of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under
-the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-OF ANY KIND, either express or implied. See the License for the specific language
-governing permissions and limitations under the License.
-*/
-
-
-/**********************************************************************************
- * 
- * Test the Browser-side implementation of the crypto node module.
- * 
- *********************************************************************************/
-
-/* eslint no-unused-vars: "off" */
- const crypto = require('../src/web/crypto.js');
- 
-describe('Crypto', function() {
-
-    it('Should load the module', async function() {
-        expect(1).toBe(1);
-    });
-
-
-});
